Animations and some backend itnegrations with menu listing

+++ database/migrations/2022_04_15_012856_create_user_payment_methods_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('user_payment_methods', function (Blueprint $table) {
             $table->integer('PK_FK_userID');
             $table->tinyInteger('PK_FK_paymentCode');
-            $table->integer('cardNo');
+            $table->bigInteger('cardNo');
             $table->float('balance');
             
             $table->primary(['PK_FK_paymentCode', 'PK_FK_userID']);
